Disabled sorting on network address
There are two things here:
1) It may not make sense to sort assigned subnets by their network
address.  Once there is sensible logic used to determine adjacency and
inclusion , the relationships defined by that logic would be a much
better choice for sorting assigned subnets.

2) SQLAlchemy doesn't like dealing with property(). Hybrid properties
may be a way to work around that while not exposing the private,
SQLAlchemy-mapped properties to the consumers of the AssignedSubnet
class.